Placeholder test for connection pool idle check
