+++ crates/holochain/src/core/workflow/validation_receipt_workflow/tests.rs
@@ -1,4 +1,4 @@
-use crate::sweettest::*;
+use crate::sweettest::{inline_zome_defs::simple_crud_zome, *};
 use crate::test_utils::consistency_10s;
 use hdk::prelude::*;
 use holo_hash::DhtOpHash;
@@ -7,22 +7,6 @@ use holochain_sqlite::prelude::*;
 use holochain_state::prelude::*;
 use holochain_types::prelude::*;
 use rusqlite::Transaction;
-
-fn simple_crud_zome() -> InlineZome {
-    let entry_def = EntryDef::default_with_id("entrydef");
-
-    InlineZome::new_unique(vec![entry_def.clone()])
-        .callback("create", move |api, ()| {
-            let entry_def_id: EntryDefId = entry_def.id.clone();
-            let entry = Entry::app(().try_into().unwrap()).unwrap();
-            let hash = api.create(EntryWithDefId::new(entry_def_id, entry))?;
-            Ok(hash)
-        })
-        .callback("read", |api, hash: HeaderHash| {
-            api.get(GetInput::new(hash.into(), GetOptions::default()))
-                .map_err(Into::into)
-        })
-}
 
 #[tokio::test(flavor = "multi_thread")]
 async fn test_validation_receipt() {
